+++ bot/src/main/resources/application.yml
@@ -45,7 +45,7 @@ kafka:
     enable-auto-commit: false
     concurrency: 1
   producer:
-    bootstrap-servers: http://localhost:9092
+    bootstrap-servers: localhost:9092
     client-id: bot
     acks-mode: 0
     delivery-timeout: 60s
