+++ scrapper/src/main/java/edu/java/scrapper/services/ScrapperQueueProducer.java
@@ -1,7 +1,5 @@
 package edu.java.scrapper.services;
 
-import edu.java.scrapper.clients.bot.BotClientImpl;
-import edu.java.scrapper.configuration.ApplicationConfig;
 import edu.java.scrapper.dtos.LinkUpdateDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,18 +14,14 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class ScrapperQueueProducer {
     private final KafkaTemplate<String, LinkUpdateDto> kafkaProducer;
-    private final BotClientImpl botClient;
-    private final ApplicationConfig applicationConfig;
 
-    public void send(LinkUpdateDto linkUpdateDto) {
-        if (applicationConfig.useQueue()) {
-            Message<LinkUpdateDto> message = MessageBuilder
-                .withPayload(linkUpdateDto)
-                .setHeader(KafkaHeaders.TOPIC, applicationConfig.kafkaTopicName())
-                .build();
-            kafkaProducer.send(message);
-        } else {
-            botClient.sendMessage(linkUpdateDto);
-        }
+    public void send(LinkUpdateDto linkUpdateDto, String topicName) {
+        log.info(linkUpdateDto.toString());
+        log.info(topicName);
+        Message<LinkUpdateDto> message = MessageBuilder
+            .withPayload(linkUpdateDto)
+            .setHeader(KafkaHeaders.TOPIC, topicName)
+            .build();
+        kafkaProducer.send(message);
     }
 }
